it compiles. also split player and link headers.
