+++ src/d/actor/d_a_item.cpp
@@ -6,7 +6,7 @@
 #include "d/d_item.h"
 #include "d/d_item_data.h"
 #include "d/actor/d_a_item.h"
-#include "d/actor/d_a_player.h"
+#include "d/actor/daPy_lk.h"
 #include "d/d_com_inf_game.h"
 #include "d/d_procname.h"
 #include "dolphin/types.h"
