+++ src/d/actor/d_a_player.cpp
@@ -3,7 +3,7 @@
 // Translation Unit: d_a_player.cpp
 //
 
-#include "d_a_player.h"
+#include "daPy_lk.h"
 #include "dolphin/types.h"
 
 /* 801028FC-80102940       .text changePlayer__9daPy_py_cFP10fopAc_ac_c */
